ok trying out better commit comments. This should be line 1
And this should be in a new paragraph with a blank line in between this and the first line.
and the rest of this should be on a new line but in the same paragraph.
Finished most of the organization and markdown.
Finished most of the EDA, still want to do more with pre- and post-covid data
Created one simple map with folium.